setup dashboard auth
