+++ app/auth/confirm/route.js
@@ -7,7 +7,7 @@ export async function GET(request) {
   const { searchParams } = new URL(request.url)
   const token_hash = searchParams.get('token_hash')
   const type = searchParams.get('type')
-  const next = searchParams.get('next') ?? '/private'
+  const next = searchParams.get('next') ?? '/dashboard'
 
   const redirectTo = request.nextUrl.clone()
   redirectTo.pathname = next
