+++ app/dashboard/work/page.jsx
@@ -1,0 +1,7 @@
+export default function Work() {
+    return <>
+        <h2>Ini Work dashboard</h2>
+    
+    
+    </>
+}
