+++ app/login/actions.js
@@ -21,8 +21,8 @@ export async function login(formData) {
     redirect('/error')
   }
 
-  revalidatePath('/private', 'layout')
-  redirect('/private')
+  revalidatePath('/dashboard', 'layout')
+  redirect('/dashboard')
 }
 
 export async function signup(formData) {
@@ -42,5 +42,5 @@ export async function signup(formData) {
   }
 
   revalidatePath('/', 'layout')
-  redirect('/')
+  redirect('/ ')
 }
