+++ app/login/page.jsx
@@ -7,14 +7,15 @@ export default async function LoginPage() {
 
   const { data } = await supabase.auth.getUser();
   if (data?.user) {
-    redirect("/private");
+    redirect("/dashboard");
   }
+
   return (
     <form>
       <label htmlFor="email">Email:</label>
-      <input id="email" name="email" type="email" required />
+      <input id="email" name="email" type="email" autoComplete="username" required />
       <label htmlFor="password">Password:</label>
-      <input id="password" name="password" type="password" required />
+      <input id="password" name="password" type="password" autoComplete="current-password" required />
       <button formAction={login}>Log in</button>
       <button formAction={signup}>Sign up</button>
     </form>
