+++ package.json
@@ -16,6 +16,8 @@
     "react-dom": "^18"
   },
   "devDependencies": {
+    "eslint": "8.57.0",
+    "eslint-config-next": "14.2.4",
     "postcss": "^8",
     "tailwindcss": "^3.4.1"
   }
